tests/wpt/web-platform-tests/infrastructure/expected-fail/precondition-in-setup.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Precondition in setup</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup(() => {
  assert_precondition(false);
});
</script>
